Version 0.4
Fixed highlighting after new line if '<' is present in a block of code. It's still a bug in oneliners, because UDL doesn't support a whitespace as a delimeter(in any shape such as regex, unicode or html entity). If there is such an option, oneliners with '<' will look pretty too.

New property keyword: 'focus'